requirements.txt
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.15.0
cryptography>=41.0.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0
requests>=2.31.0
psutil>=5.9.0
tqdm>=4.65.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0
pytest-asyncio>=0.21.0
coverage>=7.2.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.0
safety>=2.3.0

# GUI dependencies (optional)
PyQt5>=5.15.0; platform_system=="Windows" or platform_system=="Linux"
kivy>=2.2.0; platform_system=="Android"

# Platform-specific dependencies
pywin32>=306; platform_system=="Windows"
python-pam>=2.0.2; platform_system=="Linux"